SPLICE-1330 Speeding up Join Planning
