+++ db-engine/src/main/java/com/splicemachine/db/impl/ast/ColumnUtils.java
@@ -34,13 +34,9 @@ package com.splicemachine.db.impl.ast;
 import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.db.iapi.sql.compile.Visitable;
 import com.splicemachine.db.impl.sql.compile.*;
-import org.apache.commons.lang3.tuple.Pair;
 import org.spark_project.guava.collect.Iterables;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * User: pjt
@@ -56,80 +52,6 @@ public class ColumnUtils {
         public boolean apply(ColumnReference columnReference) {
             return columnReference.getCorrelated();
         }
-    }
-
-    /**
-     * For a given ResultColumnList, return a map from
-     * [resultSetNumber, virtualColumnId] => ResultColumn
-     * where there is one entry for each ResultColumn down the chain of reference to
-     * its source column on a table. This allows translation from a column reference at
-     * any node below into the ResultColumn projected from the passed ResultColumnList.
-     */
-    public static Map<Pair<Integer, Integer>, ResultColumn> rsnChainMap(ResultColumnList rcl)
-            throws StandardException {
-        Map<Pair<Integer, Integer>, ResultColumn> chain = new HashMap<Pair<Integer, Integer>, ResultColumn>();
-        List<ResultColumn> cols = RSUtils.collectNodes(rcl, ResultColumn.class);
-
-        for (ResultColumn rc : cols) {
-            Pair<Integer, Integer> top = RSCoordinate(rc);
-            chain.put(top, rc);
-            for (Pair<Integer, Integer> link : rsnChain(rc)) {
-                chain.put(link, rc);
-            }
-        }
-
-        return chain;
-    }
-
-    /**
-     * For a given ResultColumn, return a list of integer pairs [resultSetNumber, virtualColumnId]
-     * for its source column, and its source's source column, and so on down to the bottom: a source
-     * column on a table.
-     */
-    public static List<Pair<Integer, Integer>> rsnChain(ResultColumn rc)
-            throws StandardException {
-        List<Pair<Integer, Integer>> chain = new ArrayList<Pair<Integer, Integer>>();
-
-        ValueNode expression = rc.getExpression();
-        while (expression != null) {
-            if (expression instanceof VirtualColumnNode) {
-                ResultColumn sc = ((VirtualColumnNode) expression).getSourceColumn();
-                chain.add(RSCoordinate(sc));
-                expression = sc.getExpression();
-            } else if (expression instanceof ColumnReference) {
-                ResultColumn sc = ((ColumnReference) expression).getSource();
-                if (sc != null) { // A ColumnReference can be sourceless…
-                    chain.add(RSCoordinate(sc));
-                    expression = sc.getExpression();
-                } else {
-                    expression = null;
-                }
-            } else if (expression instanceof CastNode) {
-                expression = ((CastNode) expression).getCastOperand();
-            }
-            else {
-                expression = null;
-            }
-        }
-
-        return chain;
-    }
-
-    static Pair<Integer, Integer> RSCoordinate(ResultColumn rc) {
-
-        ResultColumn resultColumn = rc;
-        ValueNode vn = rc.getExpression();
-        if (vn instanceof CastNode) {
-            ValueNode castOperand=((CastNode)vn).getCastOperand();
-            if(castOperand instanceof VirtualColumnNode){
-                VirtualColumnNode vcn=(VirtualColumnNode)castOperand;
-                resultColumn=vcn.getSourceColumn();
-            }else if(castOperand instanceof ColumnReference){
-                ColumnReference cr = (ColumnReference)castOperand;
-                resultColumn = cr.getSourceResultColumn();
-            }
-        }
-        return Pair.of(resultColumn.getResultSetNumber(), resultColumn.getVirtualColumnId());
     }
 
     /**
