+++ db-engine/src/main/java/com/splicemachine/db/impl/ast/JoinConditionVisitor.java
@@ -32,28 +32,21 @@
 package com.splicemachine.db.impl.ast;
 
 
+import com.carrotsearch.hppc.LongLongOpenHashMap;
+import com.splicemachine.db.iapi.services.context.ContextService;
+import com.splicemachine.db.iapi.sql.compile.*;
+import com.splicemachine.db.iapi.sql.conn.ConnectionUtil;
+import com.splicemachine.db.iapi.sql.conn.LanguageConnectionContext;
+import com.splicemachine.db.iapi.util.JBitSet;
+import com.splicemachine.db.impl.sql.compile.*;
 import org.spark_project.guava.base.Function;
 import com.splicemachine.db.iapi.error.StandardException;
-import com.splicemachine.db.iapi.sql.compile.AccessPath;
-import com.splicemachine.db.iapi.sql.compile.Optimizable;
-import com.splicemachine.db.impl.sql.compile.ColumnReference;
-import com.splicemachine.db.impl.sql.compile.FromBaseTable;
-import com.splicemachine.db.impl.sql.compile.HalfOuterJoinNode;
-import com.splicemachine.db.impl.sql.compile.IndexToBaseRowNode;
-import com.splicemachine.db.impl.sql.compile.JoinNode;
-import com.splicemachine.db.impl.sql.compile.Predicate;
-import com.splicemachine.db.impl.sql.compile.PredicateList;
-import com.splicemachine.db.impl.sql.compile.ProjectRestrictNode;
-import com.splicemachine.db.impl.sql.compile.ResultColumn;
-import com.splicemachine.db.impl.sql.compile.ResultColumnList;
-import com.splicemachine.db.impl.sql.compile.ResultSetNode;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.log4j.Logger;
 import org.spark_project.guava.base.Predicates;
-import org.spark_project.guava.collect.Collections2;
-import org.spark_project.guava.collect.Iterables;
-import org.spark_project.guava.collect.Lists;
-import org.spark_project.guava.collect.Sets;
+import org.spark_project.guava.collect.*;
+import org.spark_project.guava.primitives.Ints;
+import java.sql.SQLException;
 import java.util.*;
 
 
@@ -78,11 +71,37 @@ import java.util.*;
 public class JoinConditionVisitor extends AbstractSpliceVisitor {
 
     private static Logger LOG = Logger.getLogger(JoinConditionVisitor.class);
+    private LongLongOpenHashMap joinChainMap;
+
+    private void initializeMap(QueryTreeNode v)  throws StandardException {
+        if (joinChainMap == null)
+            joinChainMap = v.childParentMap();
+    }
+
+    @Override
+    public Visitable visit(ProjectRestrictNode node) throws StandardException {
+        if (node.hasSubqueries())
+            initializeMap(node);
+        return super.visit(node);
+    }
+
+    @Override
+    public UnionNode visit(UnionNode node) throws StandardException {
+        initializeMap(node);
+        return node;
+    }
+
+    @Override
+    public IntersectOrExceptNode visit(IntersectOrExceptNode node) throws StandardException {
+        initializeMap(node);
+        return node;
+    }
 
     @Override
     public JoinNode visit(JoinNode j) throws StandardException {
         if (LOG.isDebugEnabled())
             LOG.debug(String.format("visit joinNode=%s",j));
+        initializeMap(j);
         AccessPath ap = ((Optimizable) j.getRightResultSet()).getTrulyTheBestAccessPath();
         if (RSUtils.isHashableJoin(ap)){
             return pullUpPreds(j, ap);
@@ -142,7 +161,7 @@ public class JoinConditionVisitor extends AbstractSpliceVisitor {
                     toPullUp.addAll(c);
             }
         }
-
+       // joinChainMap = j.rsnChainMapV2();
         for (Predicate p: toPullUp){
             p = updatePredColRefsToNode(p, j);
             j.addOptPredicate(p);
@@ -151,6 +170,16 @@ public class JoinConditionVisitor extends AbstractSpliceVisitor {
                         PredicateUtils.predToString.apply(p), j.getResultSetNumber()));
             }
         }
+
+        Set<Predicate> equiJoinPreds =
+                Sets.filter(Sets.newLinkedHashSet(PredicateUtils.PLtoList(j.joinPredicates)),
+                        PredicateUtils.isEquiJoinPred);
+
+        Pair<List<Integer>, List<Integer>> indices = findHashIndices(j, equiJoinPreds);
+        if (LOG.isDebugEnabled())
+            LOG.debug(String.format("Hash key indices found for Join n=%s: %s", j.getResultSetNumber(), indices));
+        j.leftHashKeys = Ints.toArray(indices.getLeft());
+        j.rightHashKeys = Ints.toArray(indices.getRight());
 
         return j;
     }
@@ -254,6 +283,8 @@ public class JoinConditionVisitor extends AbstractSpliceVisitor {
                                          predOfInterest));
         }
 
+      //  joinChainMap = j.rsnChainMapV2();
+
         for (Predicate p: joinPreds) {
         	updatePredColRefsToNode(p, j.getLeftResultSet());
         }
@@ -270,7 +301,7 @@ public class JoinConditionVisitor extends AbstractSpliceVisitor {
                         new Function<ColumnReference, Integer>() {
                             @Override
                             public Integer apply(ColumnReference cr) {
-                                return ColumnUtils.RSCoordinate(cr.getSource()).getLeft();
+                                return (int) (cr.getSource().getCoordinates() >> 32);
                             }
                         }));
     }
@@ -301,18 +332,260 @@ public class JoinConditionVisitor extends AbstractSpliceVisitor {
             throws StandardException {
     	if (LOG.isDebugEnabled())
     		LOG.debug(String.format("updatePredColRefsToNode predicate=%s, resultSetNode=%s",p,n));
-    	
-        ResultColumnList rcl = n.getResultColumns();
-        Map<Pair<Integer,Integer>, ResultColumn> chain = ColumnUtils.rsnChainMap(rcl);
         List<ColumnReference> predCRs = RSUtils.collectNodes(p, ColumnReference.class);
         for (ColumnReference cr: predCRs){
+            boolean ignore = false;
             ResultColumn rc = cr.getSource();
-            Pair<Integer,Integer> rsnAndCol = ColumnUtils.RSCoordinate(rc);
-            if (chain.containsKey(rsnAndCol)){
-                cr.setSource(chain.get(rsnAndCol));
+            long rsnAndCol = rc.getCoordinates();
+            if (!joinChainMap.containsKey(rsnAndCol)) {
+                continue; // huh?
             }
+            while (  (int)(rsnAndCol>>32) != n.getResultSetNumber()) {
+                System.out.println(String.format("ResultColumn(%d,%d)",(int)(rsnAndCol>>32),(int)(rsnAndCol)));
+                if (!joinChainMap.containsKey(rsnAndCol)) {
+                    ignore = true;
+                    break;
+                }
+                rsnAndCol = joinChainMap.get(rsnAndCol);
+            }
+            if (!ignore)
+                cr.setSource(n.getResultColumns().elementAt( ((int) rsnAndCol)-1 ));
         }
         return p;
     }
+    public Integer translateToIndexOnList(ColumnReference cr,
+                                                 ResultColumnList rcl,
+                                                 ValueNode operand,
+                                                 ResultSetNode resultSetNode,
+                                                 BinaryRelationalOperatorNode brop,
+                                                 JoinNode  joinNode)
+            throws StandardException{
+        boolean hasSubqueryNode = false;
+        if (cr.getSource().getExpression() instanceof VirtualColumnNode) {
+            VirtualColumnNode n = (VirtualColumnNode) cr.getSource().getExpression();
+            if (n.getSourceResultSet() instanceof FromSubquery) {
+                hasSubqueryNode = true;
+            }
+        }
 
+        if (operand instanceof ColumnReference) {
+            long colCoord = (cr.getSourceResultColumn() != null && !hasSubqueryNode ? cr.getSourceResultColumn() : cr.getOrigSourceResultColumn()).getCoordinates();
+            while (  (int)(colCoord>>32) != resultSetNode.getResultSetNumber()) {
+                //.out.println(String.format("ResultColumn(%d,%d)",(int)(colCoord>>32),(int)(colCoord)));
+                if (!joinChainMap.containsKey(colCoord)) {
+                    System.out.println(String.format("Error -> (%d,%d)",(int) (colCoord >> 32), (int) colCoord));
+                    throw new RuntimeException(String.format("Unable to find ColRef %s in RCL %s", cr, rcl));
+                }
+                colCoord = joinChainMap.get(colCoord);
+            }
+            System.out.println(String.format("Resolved(%d,%d) at (%d)",(int)(colCoord>>32),(int)(colCoord),resultSetNode.getResultSetNumber()));
+            return rcl.elementAt( ((int) colCoord)-1).getVirtualColumnId() - 1; // translate from 1-based to 0-based index
+        }
+        else {
+            // Add a result column to the child result set
+            ResultColumn resultColumn = addResultColumnToChild(cr, operand, resultSetNode, rcl);
+
+            //Change operand's column reference to the column that has just added to child result set
+            resultColumn = resetColumnReference(resultColumn,operand, brop, joinNode, resultSetNode);
+
+            // Add a column to the result column list of the join node
+            rebuildJoinNodeRCL(joinNode, resultSetNode, resultColumn);
+
+            // Hash column position is the end of result column list of child result set
+            return rcl.size()-1;
+        }
+    }
+
+    private static ResultColumn resetColumnReference(ResultColumn resultColumn,
+                                                     ValueNode operand,
+                                                     BinaryRelationalOperatorNode brop,
+                                                     JoinNode joinNode,
+                                                     ResultSetNode resultSetNode) throws StandardException {
+
+        try {
+            LanguageConnectionContext lcc = ConnectionUtil.getCurrentLCC();
+            NodeFactory nodeFactory = lcc.getLanguageConnectionFactory().
+                    getNodeFactory();
+
+            ColumnReference generatedRef = (ColumnReference) nodeFactory.getNode(
+                    C_NodeTypes.COLUMN_REFERENCE,
+                    resultColumn.getName(),
+                    null,
+                    ContextService.getService().getCurrentContextManager());
+            VirtualColumnNode vnode = (VirtualColumnNode) nodeFactory.getNode(C_NodeTypes.VIRTUAL_COLUMN_NODE,
+                    resultSetNode, // source result set.
+                    resultColumn,
+                    resultSetNode.getResultColumns().size(),
+                    ContextService.getService().getCurrentContextManager());
+
+            resultColumn = (ResultColumn) nodeFactory.getNode(C_NodeTypes.RESULT_COLUMN,
+                    resultColumn.getName(),
+                    vnode,
+                    ContextService.getService().getCurrentContextManager());
+            resultColumn.markGenerated();
+            resultColumn.setResultSetNumber(joinNode.getResultSetNumber());
+            generatedRef.setSource(resultColumn);
+            if (brop.getLeftOperand() == operand) {
+                generatedRef.setTableNumber(brop.getLeftOperand().getHashableJoinColumnReference().getTableNumber());
+                brop.setLeftOperand(generatedRef);
+            }
+            else {
+                generatedRef.setTableNumber(brop.getRightOperand().getHashableJoinColumnReference().getTableNumber());
+                brop.setRightOperand(generatedRef);
+            }
+            return resultColumn;
+        }
+        catch (SQLException e) {
+            throw StandardException.newException(e.getSQLState());
+        }
+    }
+
+    private static ResultColumn addResultColumnToChild(ColumnReference cr,
+                                                       ValueNode operand,
+                                                       ResultSetNode resultSetNode,
+                                                       ResultColumnList rcl) throws StandardException{
+        try {
+            LanguageConnectionContext lcc = ConnectionUtil.getCurrentLCC();
+            NodeFactory nodeFactory = lcc.getLanguageConnectionFactory().
+                    getNodeFactory();
+
+            // Add a result column and return virtual column Id
+            ResultColumn rc = cr.getSource();
+            assert rc.getExpression() instanceof VirtualColumnNode;
+            VirtualColumnNode vn = (VirtualColumnNode) rc.getExpression();
+            rc = vn.getSourceColumn();
+
+            // construct a result column and add to result column list of child result set
+            ColumnReference generatedRef = (ColumnReference) nodeFactory.getNode(
+                    C_NodeTypes.COLUMN_REFERENCE,
+                    rc.getName(),
+                    null,
+                    ContextService.getService().getCurrentContextManager());
+
+            assert rc.getExpression() instanceof VirtualColumnNode;
+            vn = (VirtualColumnNode) rc.getExpression();
+            generatedRef.setSource(vn.getSourceColumn());
+            generatedRef.setTableNumber(operand.getHashableJoinColumnReference().getTableNumber());
+            operand.setHashableJoinColumnReference(generatedRef);
+
+            ResultColumn resultColumn =
+                    (ResultColumn) nodeFactory.getNode(C_NodeTypes.RESULT_COLUMN,
+                            generatedRef.getColumnName(),
+                            operand,
+                            ContextService.getService().getCurrentContextManager());
+            resultColumn.markGenerated();
+            resultColumn.setResultSetNumber(resultSetNode.getResultSetNumber());
+            resultColumn.setVirtualColumnId(rcl.size());
+            rcl.addResultColumn(resultColumn);
+
+            return resultColumn;
+        }
+        catch (SQLException e) {
+            throw StandardException.newException(e.getSQLState());
+        }
+    }
+    public Pair<List<Integer>, List<Integer>> findHashIndices(final JoinNode node, Collection<Predicate> equiJoinPreds)
+            throws StandardException {
+        List<Integer> leftIndices = Lists.newArrayListWithCapacity(equiJoinPreds.size());
+        List<Integer> rightIndices = Lists.newArrayListWithCapacity(equiJoinPreds.size());
+        org.spark_project.guava.base.Predicate<ResultColumn> isLeftRef = RSUtils.pointsTo(node.getLeftResultSet());
+        ResultColumnList leftRCL = node.getLeftResultSet().getResultColumns();
+        ResultColumnList rightRCL = node.getRightResultSet().getResultColumns();
+
+        for (Predicate p : equiJoinPreds) {
+            AndNode andNode = p.getAndNode();
+            assert andNode.getLeftOperand() instanceof BinaryRelationalOperatorNode;
+            BinaryRelationalOperatorNode brop = (BinaryRelationalOperatorNode)andNode.getLeftOperand();
+            ValueNode leftOperand = brop.getLeftOperand();
+            ValueNode rightOperand = brop.getRightOperand();
+            List<ColumnReference> lcr = RSUtils.collectNodes(leftOperand, ColumnReference.class);
+            List<ColumnReference> rcr = RSUtils.collectNodes(rightOperand, ColumnReference.class);
+
+            for (ColumnReference cr : lcr) {
+                if (isLeftRef.apply(cr.getSourceResultColumn()!=null?cr.getSourceResultColumn():cr.getOrigSourceResultColumn())) {
+                    leftIndices.add(translateToIndexOnList(cr, leftRCL, leftOperand, node.getLeftResultSet(), brop, node));
+                } else {
+                    rightIndices.add(translateToIndexOnList(cr, rightRCL, leftOperand, node.getRightResultSet(), brop, node));
+                }
+            }
+
+            for (ColumnReference cr : rcr) {
+                if (isLeftRef.apply(cr.getSourceResultColumn()!=null?cr.getSourceResultColumn():cr.getOrigSourceResultColumn())) {
+                    leftIndices.add(translateToIndexOnList(cr, leftRCL, rightOperand, node.getLeftResultSet(), brop, node));
+                } else {
+                    rightIndices.add(translateToIndexOnList(cr, rightRCL, rightOperand, node.getRightResultSet(), brop, node));
+                }
+            }
+        }
+
+        return Pair.of(leftIndices, rightIndices);
+    }
+
+    private static void rebuildJoinNodeRCL(JoinNode node,
+                                           ResultSetNode child,
+                                           ResultColumn rc) throws StandardException{
+
+        boolean isLeft = node.getLeftResultSet() == child ? true : false;
+        ResultColumnList rcl = node.getResultColumns();
+
+        if (isLeft) {
+            int size = rcl.size();
+            JBitSet leftReferencedTableMap=node.getLeftResultSet().getReferencedTableMap();
+            // make a copy of result column list
+            ResultColumnList temp = new ResultColumnList();
+            for (int i = 0; i < size; ++i) {
+                temp.addResultColumn(rcl.elementAt(0));
+                rcl.removeElementAt(0);
+            }
+
+            // copy result columns that reference to left result set
+            Set<ResultSetNode>  leftResultSetNodeSet = getResultSetNodes(node.getLeftResultSet());
+            Iterator<ResultColumn> iter=temp.iterator();
+            while(iter.hasNext()){
+                ResultColumn resultColumn = iter.next();
+                if (fromResultSets(leftResultSetNodeSet, resultColumn)) {
+                    rcl.addResultColumn(resultColumn);
+                    iter.remove();
+                }else break;
+            }
+
+            // Add a new column to join result column list
+            rcl.addResultColumn(rc);
+
+            // copy columns that reference to the right result set
+            while(temp.size() > 0) {
+                rcl.addResultColumn(temp.elementAt(0));
+                temp.removeElementAt(0);
+            }
+        }
+        else {
+            rcl.addResultColumn(rc);
+        }
+    }
+
+    private static Set<ResultSetNode> getResultSetNodes(ResultSetNode node) throws StandardException {
+        List<ResultSetNode> resultSetNodeList = RSUtils.collectNodes(node, ResultSetNode.class);
+        Set<ResultSetNode>  resultSetNodeSet = new HashSet<>();
+        for(ResultSetNode resultSetNode : resultSetNodeList) {
+            resultSetNodeSet.add(resultSetNode);
+        }
+        return resultSetNodeSet;
+    }
+
+    private static boolean fromResultSets(Set<ResultSetNode>  resultSetNodeSet, ResultColumn resultColumn) {
+        ValueNode v = resultColumn.getExpression();
+        if (v instanceof VirtualColumnNode) {
+            ResultSetNode sourceResultSet = ((VirtualColumnNode) v).getSourceResultSet();
+            if (resultSetNodeSet.contains(sourceResultSet))
+                return true;
+        }
+        return false;
+    }
+
+
+    @Override
+    public boolean isPostOrder() {
+        // Default to PostOrder traversal, i.e. bottom-up
+        return false;
+    }
 }
